+++ lib/model/discount.dart
@@ -3,7 +3,7 @@ part of 'model.dart';
 class Discount extends Equatable {
   final int id;
   final String name;
-  final double amount;
+  final int amount;
   final int storeId;
 
   const Discount({
@@ -16,7 +16,7 @@ class Discount extends Equatable {
   factory Discount.fromJson(Map<String, dynamic> json) => Discount(
         id: json['id'] as int,
         name: json['name'] as String,
-        amount: json['amount'] as double,
+        amount: json['amount'] as int,
         storeId: json['store_id'] as int,
       );
